Upload testpoint map on behalf of BBC
This file is (C) British Broadcasting Corporation and released on behalf of the BBC under the same license as the rest of the contents of this repository. I am uploading on behalf of the BBC.

The test point map is useful in conjunction with the schematic for knowing where the test signals are routed to.